Integrate sign up, sign in, sign out with Firebase auth and Firestore
