+++ frontend/navigation/PublicStack.js
@@ -1,0 +1,16 @@
+import React from "react";
+import { createNativeStackNavigator } from "@react-navigation/native-stack";
+import SignInScreen from "../screens/SignInScreen";
+import SignUpScreen from "../screens/SignUpScreen";
+
+const Stack = createNativeStackNavigator();
+
+export default function PublicStack() {
+  return (
+    <Stack.Navigator screenOptions={{ headerShown: false }}>
+      <Stack.Screen name="SignIn" component={SignInScreen} />
+      <Stack.Screen name="SignUp" component={SignUpScreen} />
+    </Stack.Navigator>
+  );
+}
+// This PublicStack component is used for the public routes of the application
